fix: fix text box zooming
