+++ frontend/src/features/authoring/Canvas/componentResolver.jsx
@@ -14,7 +14,7 @@ import TextComponent from "../components/TextComponent";
  *   componentResolver(component, index, selectElement)
  * )}
  */
-export default function componentResolver(component, index, onClick, isEditing) {
+export default function componentResolver(component, index, onClick) {
   switch (component.type) {
     // ADD NEW COMPONENT TYPES HERE
     case "BUTTON":
@@ -24,7 +24,6 @@ export default function componentResolver(component, index, onClick, isEditing) 
           id={index}
           onClick={onClick}
           component={component}
-          isEditing={isEditing}
         />
       );
     case "RESET_BUTTON":
@@ -34,7 +33,6 @@ export default function componentResolver(component, index, onClick, isEditing) 
           id={index}
           onClick={onClick}
           component={component}
-          isEditing={isEditing}
         />
       );
     case "SPEECHTEXT":
