+++ frontend/src/features/authoring/components/TextComponent.jsx
@@ -14,8 +14,10 @@ import useStyles from "./components.styles";
  *    onDoubleClick={handleDoubleClick}
  * />
  */
-export default function TextComponent({ id, onClick, component }) {
+export default function TextComponent({ id, onClick, component, zoomLevel }) {
   const styles = useStyles(component);
+  const userDefinedWidth = component.width;
+  const userDefinedHeight = component.height;
 
   const context = useContext(AuthoringToolContext);
   const textRef = context?.propertiesRefs?.text;
@@ -26,7 +28,11 @@ export default function TextComponent({ id, onClick, component }) {
       id={id}
       onClick={onClick}
       onDoubleClick={() => textRef?.current?.focus()}
-      style={{ zIndex: component?.zPosition || "0" }}
+      style={{
+        zIndex: component?.zPosition || "0",
+        width: `${userDefinedWidth * zoomLevel}%`,
+        height: `${userDefinedHeight * zoomLevel}%`,
+      }}
     >
       {component.text}
     </Typography>
