+++ frontend/src/features/playScenario/PlayScenarioCanvas.jsx
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import ResetConfirmationModal from "./modals/ResetConfirmationModal";
 import componentResolver from "./componentResolver";
 
@@ -16,6 +16,20 @@ export default function PlayScenarioCanvas({
   setRemoveFlags,
 }) {
   const [isModalOpen, setIsModalOpen] = useState(false);
+
+  const [zoomLevel, setZoomLevel] = useState(1);
+
+  useEffect(() => {
+    const handleResize = () => {
+      setZoomLevel(window.devicePixelRatio);
+    };
+
+    window.addEventListener("resize", handleResize);
+
+    return () => {
+      window.removeEventListener("resize", handleResize);
+    };
+  }, []);
 
   const handleResetClick = () => {
     setIsModalOpen(true);
@@ -75,7 +89,7 @@ export default function PlayScenarioCanvas({
             break;
         }
 
-        return componentResolver(component, index, action);
+        return componentResolver(component, index, action, zoomLevel);
       })}
       <ResetConfirmationModal
         isOpen={isModalOpen}
